requirements.txt
streamlit==1.28.1
google-generativeai==0.3.2
pdfminer.six==20221105
python-docx==0.8.11
pandas==2.1.3
python-dotenv==1.0.0
spacy==3.7.2
textblob==0.17.1
nest-asyncio==1.5.8
openpyxl==3.1.2
asyncio==3.4.3
aiofiles==23.2.1
tqdm==4.66.1
concurrent-futures==3.1.1